*split MergeScale.cpp into MergeConvolutionAndScale.cpp and MergeInnerProductAndScale.cpp.
