+++ prj/vs2022/CvtCore.vcxproj
@@ -28,9 +28,10 @@
     <ClCompile Include="..\..\src\Cvt\InferenceEngine\InferenceEngine.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeActivation.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeConvolution.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeConvolutionAndScale.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeInnerProductAndScale.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantized.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantizedSqueezeExcitation.cpp" />
-    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeScale.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeSqueezeExcitation.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\Optimizer.cpp" />
     <ClInclude Include="..\..\src\Cvt\Common\Params.h" />
